# docs/swagger.yaml
basePath: /api
definitions:
  controllers.ChangePasswordRequest:
    properties:
      confirm_password:
        example: password_baru123
        type: string
      new_password:
        example: password_baru123
        minLength: 8
        type: string
      old_password:
        example: password_lama123
        type: string
    required:
    - confirm_password
    - new_password
    - old_password
    type: object
...
  controllers.LoginRequest:
    properties:
      nrp:
        example: "12345"
        type: string
      password:
        example: password123
        type: string
    required:
    - nrp
    - password
    type: object
  controllers.UpdateProfileRequest:
    properties:
      nama_lengkap:
        example: NAMA SAYA
        type: string
      nrp:
        example: "12345"
        type: string
      pangkat:
        example: BRIPKA
        type: string
    required:
    - nama_lengkap
    - nrp
    - pangkat
    type: object
...
  models.AuditLog:
    properties:
      aksi:
        type: string
      detail:
        type: string
      id:
        type: integer
      timestamp:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: yusuf.aar@gmail.com
    name: MUHAMMAD YUSUF ABDURROHMAN
    url: https://github.com/dope-s/simdokpol-go
  description: Ini adalah dokumentasi API untuk aplikasi Sistem Informasi Manajemen
    Dokumen Kepolisian.
  license:
    name: MIT License
    url: https://github.com/dope-s/simdokpol-go/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: SIMDOKPOL API
  version: "1.0"
paths:
  /audit-logs:
    get:
      description: Mengambil seluruh riwayat aktivitas yang tercatat di sistem. Hanya
        bisa diakses oleh Super Admin.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AuditLog'
            type: array
        "500":
          description: 'Error: Gagal mengambil data log audit'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Semua Log Audit
      tags:
      - Audit Log
  /audit-logs/export:
    get:
      description: Mengunduh seluruh riwayat log audit sebagai file .xlsx. Hanya bisa
        diakses oleh Super Admin.
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: File Excel (.xlsx)
          schema:
            type: file
        "500":
          description: 'Error: Gagal membuat file ekspor'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ekspor Log Audit ke Excel
      tags:
      - Audit Log
  /backups:
    post:
      description: Membuat salinan database saat ini dan mengirimkannya sebagai file
        unduhan. Hanya bisa diakses oleh Super Admin.
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File backup database (.db)
          schema:
            type: file
        "500":
          description: 'Error: Gagal memproses backup'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Membuat Backup Database
      tags:
      - Backup & Restore
  /documents:
    get:
      description: Mengambil daftar semua surat keterangan hilang, bisa difilter berdasarkan
        status (aktif/arsip) dan query pencarian.
      parameters:
      - description: Kata Kunci Pencarian (No. Surat / Nama)
        in: query
        name: q
        type: string
      - default: active
        description: Filter status dokumen
        enum:
        - active
        - archived
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LostDocument'
            type: array
        "500":
          description: 'Error: Terjadi kesalahan pada server'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Semua Dokumen
      tags:
      - Documents
    post:
      consumes:
      - application/json
      description: Membuat surat keterangan hilang baru.
      parameters:
      - description: Data Dokumen Baru
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/controllers.DocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LostDocument'
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal membuat dokumen'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Membuat Dokumen Baru
      tags:
      - Documents
  /documents/export:
    get:
      description: Mengunduh daftar dokumen (aktif atau arsip) sebagai file .xlsx.
        Hanya bisa diakses oleh Super Admin.
      parameters:
      - description: Kata Kunci Pencarian (No. Surat / Nama)
        in: query
        name: q
        type: string
      - default: active
        description: Filter status dokumen
        enum:
        - active
        - archived
        in: query
        name: status
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: File Excel (.xlsx)
          schema:
            type: file
        "500":
          description: 'Error: Gagal membuat file ekspor'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ekspor Dokumen ke Excel
      tags:
      - Documents
  /documents/{id}:
    delete:
      description: Menghapus (soft delete) sebuah surat keterangan hilang. Hanya bisa
        diakses oleh Super Admin atau operator yang membuatnya.
      produces:
      - application/json
      parameters:
      - description: ID Dokumen
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Pesan Sukses
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Akses ditolak'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal menghapus dokumen'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Menghapus Dokumen
      tags:
      - Documents
    get:
      description: Mengambil detail satu surat keterangan hilang berdasarkan ID-nya.
        Hanya bisa diakses oleh Super Admin atau operator yang membuat dokumen tersebut.
      produces:
      - application/json
      parameters:
      - description: ID Dokumen
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LostDocument'
        "400":
          description: 'Error: ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Akses ditolak'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: Dokumen tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mendapatkan Dokumen Berdasarkan ID
      tags:
      - Documents
    put:
      consumes:
      - application/json
      description: Memperbarui data sebuah surat keterangan hilang. Hanya bisa diakses
        oleh Super Admin atau operator yang membuatnya.
      parameters:
      - description: ID Dokumen
        in: path
        name: id
        required: true
        type: integer
      - description: Data Dokumen yang Diperbarui
        in: body
        name: document
        required: true
        schema:
          $ref: '#/definitions/controllers.DocumentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LostDocument'
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Akses ditolak'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal memperbarui dokumen'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Memperbarui Dokumen
      tags:
      - Documents
  /documents/{id}/pdf:
    get:
      description: Mengunduh satu surat keterangan hilang sebagai file .pdf yang presisi.
      produces:
      - application/pdf
      parameters:
      - description: ID Dokumen
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: File PDF (.pdf)
          schema:
            type: file
        "400":
          description: 'Error: ID tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Akses ditolak'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'Error: Dokumen tidak ditemukan'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal membuat file PDF'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download Dokumen sebagai PDF
      tags:
      - Documents
  /login:
    post:
      consumes:
      - application/json
      description: Melakukan otentikasi pengguna berdasarkan NRP dan kata sandi, lalu
        mengembalikan token JWT dalam HttpOnly cookie.
      parameters:
      - description: Data Login Pengguna
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Contoh: {"message": "Login berhasil"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Contoh: {"error": "NRP dan Kata Sandi diperlukan"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'Contoh: {"error": "NRP atau kata sandi salah"}'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login Pengguna
      tags:
      - Authentication
  /notifications/expiring-documents:
    get:
      description: Mengambil daftar dokumen yang akan segera masuk masa arsip untuk
        pengguna yang sedang login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LostDocument'
            type: array
      security:
      - BearerAuth: []
      summary: Mendapatkan Notifikasi Dokumen Kedaluwarsa
      tags:
      - Dashboard & Stats
  /profile:
    put:
      consumes:
      - application/json
      description: Memperbarui data profil (Nama, NRP, Pangkat) untuk pengguna yang
        sedang login.
      parameters:
      - description: Data Profil Baru
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Pesan sukses dan data user yang diperbarui
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Terjadi kesalahan pada server'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Memperbarui Profil Pengguna
      tags:
      - Profile
  /profile/password:
    put:
      consumes:
      - application/json
      description: Mengubah kata sandi untuk pengguna yang sedang login.
      parameters:
      - description: Data Kata Sandi Lama dan Baru
        in: body
        name: passwords
        required: true
        schema:
          $ref: '#/definitions/controllers.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Pesan Sukses
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 'Error: Kata sandi lama salah'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mengubah Kata Sandi Pengguna
      tags:
      - Profile
  /restore:
    post:
      consumes:
      - multipart/form-data
      description: Memulihkan database dari file .db yang diunggah. Semua data saat
        ini akan ditimpa. Hanya bisa diakses oleh Super Admin.
      parameters:
      - description: File backup .db yang akan di-restore
        in: formData
        name: restore-file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Pesan Sukses
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal memulihkan database'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Melakukan Restore Database
      tags:
      - Backup & Restore
  /search:
    get:
      description: Mencari dokumen (aktif dan arsip) berdasarkan Nomor Surat atau Nama
        Pemohon.
      parameters:
      - description: Kata Kunci Pencarian
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LostDocument'
            type: array
        "500":
          description: 'Error: Terjadi kesalahan pada server'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Pencarian Dokumen Global
      tags:
      - Documents
...
  /setup/restore:
    post:
      consumes:
      - multipart/form-data
      description: Memulihkan database dari file .db yang diunggah. Hanya bisa diakses
        saat setup awal.
      parameters:
      - description: File backup .db yang akan di-restore
        in: formData
        name: restore-file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Pesan Sukses
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Error: Input tidak valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'Error: Aplikasi sudah dikonfigurasi'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Error: Gagal memulihkan database'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Pulihkan Database Saat Setup
      tags:
      - Setup
...
securityDefinitions:
  BearerAuth:
    description: Masukkan token JWT Anda dengan format 'Bearer {token}'.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"